--- server/.gitignore
@@ -1,4 +1,0 @@
-node_modules
-.idea/dataSources/
-.idea
-.idea/*
