+++ src/reusable.scss
@@ -894,7 +894,10 @@
 .add-color-orange {
   color: orange !important;
 }
-
+//color white
+.add-color-white {
+  color: white !important;
+}
 //width-75%
 .crop-width-75 {
   width: 75% !important;
